move pageFromQuery to util
