+++ pages/exhibitions/index.vue
@@ -40,6 +40,9 @@
   import createClient from '../../plugins/contentful';
   import ContentCard from '../../components/generic/ContentCard';
   import PaginationNav from '../../components/generic/PaginationNav';
+  import { pageFromQuery } from '../../plugins/utils';
+
+  const PER_PAGE = 20;
 
   export default {
     name: 'ExhibitionFoyer',
@@ -54,8 +57,8 @@
     },
     data() {
       return {
-        perPage: 20,
-        page: Number(this.$route.query.page || 1)
+        perPage: PER_PAGE,
+        page: null
       };
     },
     computed: {
@@ -64,21 +67,21 @@
       }
     },
     asyncData({ query, redirect, error, app, store }) {
-      const contentfulClient = createClient(query.mode);
-      const currentPage = query && query.page;
-      const limit = 20;
-      if (!currentPage) {
+      const currentPage = pageFromQuery(query.page);
+      if (currentPage === null) {
         // Redirect non-positive integer values for `page` to `page=1`
         query.page = '1';
         return redirect(app.localePath({ name: 'exhibitions', query }));
       }
+
+      const contentfulClient = createClient(query.mode);
       return contentfulClient.getEntries({
         'locale': app.i18n.isoLocale(),
         'content_type': 'exhibitionPage',
-        'skip': (Number(currentPage) - 1) * limit,
+        'skip': (currentPage - 1) * PER_PAGE,
         // TODO refactor this to use firstPublishedAt, which is not searchable and may require a custom field + webhook
         'order': '-sys.createdAt',
-        limit
+        limit: PER_PAGE
       })
         .then((response) => {
           store.commit('breadcrumb/setBreadcrumbs', [
@@ -89,7 +92,9 @@
           ]);
           return {
             exhibitions: response.items,
-            total: response.total
+            total: response.total,
+            page: currentPage,
+            perPage: PER_PAGE
           };
         })
         .catch((e) => {
